refactor: break apart tests
